mtk: connectivity: Synchronize codes for Gigaset GX4PRO alps-mp-t0.mp1.rc-V11.40.t0mp1rc.k6985v1.64_P21

+++ drivers/misc/mediatek/connectivity/wlan/core/gen2/mgmt/privacy.c
@@ -162,11 +162,12 @@ VOID secInit(IN P_ADAPTER_T prAdapter, IN UINT_8 ucNetTypeIdx)
 
 	cnmTimerInitTimer(prAdapter,
 			  &prAisSpecBssInfo->rPreauthenticationTimer,
-			  (PFN_MGMT_TIMEOUT_FUNC) rsnIndicatePmkidCand, (ULONG) NULL);
+			  (PFN_MGMT_TIMEOUT_FUNC) rsnIndicatePmkidCand, (uintptr_t) NULL);
 #endif
 #if CFG_SUPPORT_802_11W
 	cnmTimerInitTimer(prAdapter,
-			  &prAisSpecBssInfo->rSaQueryTimer, (PFN_MGMT_TIMEOUT_FUNC) rsnStartSaQueryTimer, (ULONG) NULL);
+	 &prAisSpecBssInfo->rSaQueryTimer,
+	 (PFN_MGMT_TIMEOUT_FUNC) rsnStartSaQueryTimer, (uintptr_t) NULL);
 #endif
 
 	prAisSpecBssInfo->fgCounterMeasure = FALSE;
